refactor server/index.js
